+++ cmd/root.go
@@ -4,7 +4,7 @@ import "github.com/spf13/cobra"
 
 func NewRootCmd() *cobra.Command {
 	rootCmd := &cobra.Command{
-		Use:   "bizzmod",
+		Use:   "uproc-processes",
 		Short: "Bizzmod CLI for /api/v1/external",
 	}